+++ backend/Squirrel.ConsoleApp.BL/Providers/SqlServerScripts/GetObjects.cs
@@ -1,12 +1,12 @@
-﻿namespace Squirrel.ConsoleApp.Providers.SqlServerScripts
-{
-    internal class GetObjects
-    {
-        public static string GetStoredProceduresNamesScript =>
-            @"SELECT NAME as ProcedureName FROM SYS.OBJECTS WHERE TYPE_DESC = 'SQL_STORED_PROCEDURE'";
+﻿namespace Squirrel.ConsoleApp.BL.Providers.SqlServerScripts;
 
-        public static string GetStoredProcedureDefinitionScript(string storedProcedureName) =>
-            @$"
+internal class GetObjects
+{
+    public static string GetStoredProceduresNamesScript =>
+        @"SELECT NAME as ProcedureName FROM SYS.OBJECTS WHERE TYPE_DESC = 'SQL_STORED_PROCEDURE'";
+
+    public static string GetStoredProcedureDefinitionScript(string storedProcedureName) =>
+        @$"
                 SELECT M.definition [ROUTINE_DEFINITION] 
 
                 FROM INFORMATION_SCHEMA.ROUTINES R INNER JOIN sys.sql_modules M ON M.object_id = OBJECT_ID(R.ROUTINE_NAME)
@@ -14,11 +14,11 @@
                 WHERE ROUTINE_TYPE = 'PROCEDURE' AND ROUTINE_NAME = '{storedProcedureName}'
             ";
 
-        public static string GetFunctionsNamesScript =>
-            @"SELECT NAME as FunctionName FROM SYS.OBJECTS WHERE TYPE_DESC = 'SQL_SCALAR_FUNCTION' OR TYPE_DESC = 'SQL_TABLE_VALUED_FUNCTION' OR TYPE_DESC = 'SQL_INLINE_TABLE_VALUED_FUNCTION'";
+    public static string GetFunctionsNamesScript =>
+        @"SELECT NAME as FunctionName FROM SYS.OBJECTS WHERE TYPE_DESC = 'SQL_SCALAR_FUNCTION' OR TYPE_DESC = 'SQL_TABLE_VALUED_FUNCTION' OR TYPE_DESC = 'SQL_INLINE_TABLE_VALUED_FUNCTION'";
 
-        public static string GetFunctionDefinitionScript(string functionName) =>
-            @$"
+    public static string GetFunctionDefinitionScript(string functionName) =>
+        @$"
                 SELECT M.definition [ROUTINE_DEFINITION] 
 
                 FROM INFORMATION_SCHEMA.ROUTINES R INNER JOIN sys.sql_modules M ON M.object_id = OBJECT_ID(R.ROUTINE_NAME)
@@ -26,11 +26,11 @@
                 WHERE ROUTINE_TYPE = 'FUNCTION' AND ROUTINE_NAME = '{functionName}'
             ";
 
-        public static string GetViewsNamesScript =>
-            @$"SELECT NAME as ViewName FROM SYS.OBJECTS WHERE TYPE_DESC = 'VIEW'";
+    public static string GetViewsNamesScript =>
+        @$"SELECT NAME as ViewName FROM SYS.OBJECTS WHERE TYPE_DESC = 'VIEW'";
 
-        public static string GetViewDefinitionScript(string viewName) =>
-            @$"
+    public static string GetViewDefinitionScript(string viewName) =>
+        @$"
                 SELECT M.definition [VIEW_DEFINITION] 
 
                 FROM INFORMATION_SCHEMA.VIEWS V INNER JOIN sys.sql_modules M ON M.object_id = OBJECT_ID(V.TABLE_NAME)
@@ -38,8 +38,8 @@
                 WHERE TABLE_NAME = '{viewName}'
             ";
 
-        public static string GetStoredProceduresScript =>
-            @"
+    public static string GetStoredProceduresScript =>
+        @"
                 SELECT ROUTINE_SCHEMA [Schema],
 		               ROUTINE_NAME [Name],
 		               M.definition [Definition]
@@ -49,8 +49,8 @@
                 WHERE ROUTINE_TYPE = 'PROCEDURE'
             ";
 
-        public static string GetFunctionsScript =>
-            @"
+    public static string GetFunctionsScript =>
+        @"
                 SELECT ROUTINE_SCHEMA [Schema],
 		               ROUTINE_NAME [Name],
 		               DATA_TYPE [ReturnedType],
@@ -61,13 +61,12 @@
                 WHERE ROUTINE_TYPE = 'FUNCTION'
             ";
 
-        public static string GetViewsScript =>
-            @"
+    public static string GetViewsScript =>
+        @"
                 SELECT TABLE_SCHEMA [Schema],
 		               TABLE_NAME [View],
 		               M.definition [Definition]
                 FROM INFORMATION_SCHEMA.VIEWS V
 		        INNER JOIN sys.sql_modules M ON M.object_id = OBJECT_ID(V.TABLE_NAME)
             ";
-    }
 }
