+++ backend/Squirrel.ConsoleApp/Extensions/HubConnectionExtensions.cs
@@ -16,7 +16,6 @@ public static class HubConnectionExtensions
         hubConnection.On("GetAllTablesNamesAsync", (string clientId) =>
         {
             var getActionsService = app.ApplicationServices.GetRequiredService<IGetActionsService>();
-            var test = getActionsService.GetAllTablesNamesAsync().Result;
             hubConnection.InvokeAsync("ProcessReceivedDataFromClientSide", clientId, "GetAllTablesNamesAsync", getActionsService.GetAllTablesNamesAsync().Result);
         });
