+++ backend/Squirrel.ConsoleApp/Startup.cs
@@ -6,7 +6,6 @@ using Newtonsoft.Json.Converters;
 using Squirrel.ConsoleApp.BL.Interfaces;
 using Squirrel.ConsoleApp.BL.Services;
 using Squirrel.ConsoleApp.Filters;
-using Squirrel.ConsoleApp.Services;
 using Squirrel.Core.WebAPI.Extensions;
 using Squirrel.Core.WebAPI.Validators.Project;
