+++ backend/Squirrel.SqlService/Squirrel.SqlService.BLL/Extensions/Parsing.cs
@@ -1,4 +1,4 @@
-﻿namespace Squirrel.ConsoleApp.BL.Extensions;
+﻿namespace Squirrel.SqlService.BLL.Extensions;
 
 public static class Parsing
 {
