+++ backend/Squirrel.SqlService/Squirrel.SqlService.BLL/Squirrel.SqlService.BLL.csproj
@@ -7,6 +7,8 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="AutoMapper" Version="11.0.1" />
+    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="11.0.0" />
     <PackageReference Include="DiffPlex" Version="1.7.1" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Common" Version="6.0.21" />
     <PackageReference Include="Microsoft.AspNetCore.SignalR.Core" Version="1.1.0" />
@@ -16,6 +18,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\Squirrel.AzureBlobStorage\Squirrel.AzureBlobStorage.csproj" />
+    <ProjectReference Include="..\..\Squirrel.ConsoleApp.Models\Squirrel.ConsoleApp.Models.csproj" />
     <ProjectReference Include="..\..\Squirrel.Core\Squirrel.Core.DAL\Squirrel.Core.DAL.csproj" />
     <ProjectReference Include="..\..\Squirrel.Shared\Squirrel.Shared.csproj" />
   </ItemGroup>
