+++ backend/Squirrel.SqlService/Squirrel.SqlService.WebApi/Program.cs
@@ -6,6 +6,7 @@ using Squirrel.Core.WebAPI.Extensions;
 using Squirrel.Shared.Middlewares;
 using Squirrel.SqlService.WebApi.Extensions;
 using Squirrel.SqlService.WebApi.Middlewares;
+using Squirrel.SqlService.BLL.Extensions;
 
 var builder = WebApplication.CreateBuilder(args);
 
@@ -22,7 +23,7 @@ builder.Services.AddAzureBlobStorage(builder.Configuration);
 builder.Services.AddSignalR();
 builder.Services.AddSingleton<IUserIdProvider, CustomUserIdProvider>();
 builder.Services.RegisterCustomServices(builder.Configuration);
-
+builder.Services.AddAutoMapper();
 builder.Services.AddSwaggerGen();
 
 var app = builder.Build();
